Add API endpoint for fetching JITP
This commit adds an additional endpoint to retrieve JITP for a given
CACertificate. The certificate is identified by base64 encoded SKI
resulting in an url like:

GET orgs/test/jitp/NfGQXczGpi24p2/zGmJiLI4ahs4=

which returns:

{:ok,
 %{
   "data" => %{
     "description" => "jitp description",
     "product" => "test",
     "tags" => ["test"]
   }
 }}

The CACertificateController.create was also updated to support passing
in JITP as base64 encoded json. The "product" is expected to passed as
the name of an existing product. The endpoint was added to support
testing the JITP feature in the nerves_hub_user_api library.